requirements.txt
FastAPI
uvicorn
pydantic
pytest
httpx
python-dotenv
pydantic_settings
mkdocs